requirements.txt
anyascii==0.3.3
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contractions==0.1.73
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
feedparser==6.0.12
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
idna==3.11
inexactsearch==1.0.2
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==5.4.0
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
newspaper4k==0.9.5
nltk==3.9.4
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
outcome==1.3.0.post0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.3.0
Pygments==2.20.0
PySocks==1.7.1
pyspellchecker==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.43.0
setuptools==81.0.0
sgmllib3k==1.0.0
silpa_common==0.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
textsearch==0.0.24
threadpoolctl==3.6.0
tldextract==5.3.1
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
trio==0.33.0
trio-websocket==0.12.2
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
w3lib==2.4.1
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
